Revert "handle System.OverflowException while generating coverage report"
This reverts commit edb05314a52ab749a0d6f3c5623d81a8b401c39b.

+++ src/MiniCover.Core/Hits/HitsInfo.cs
@@ -30,7 +30,7 @@ namespace MiniCover.Core.Hits
             return _valuesById.ContainsKey(id);
         }
 
-        public long GetHitCount(int id)
+        public int GetHitCount(int id)
         {
             if (!_valuesById.TryGetValue(id, out var values))
                 return 0;
@@ -48,7 +48,7 @@ namespace MiniCover.Core.Hits
 
         class HitValues
         {
-            public long HitCount { get; set; }
+            public int HitCount { get; set; }
             public HitContext[] Contexts { get; set; }
         }
     }
